+++ pkg/controllers/nodeclaim/disruption/controller.go
@@ -19,6 +19,7 @@ package disruption
 import (
 	"context"
 
+	"github.com/samber/lo"
 	"go.uber.org/multierr"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
@@ -28,6 +29,7 @@ import (
 	controllerruntime "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
@@ -35,7 +37,6 @@ import (
 	"sigs.k8s.io/karpenter/pkg/cloudprovider"
 	"sigs.k8s.io/karpenter/pkg/controllers/state"
 	operatorcontroller "sigs.k8s.io/karpenter/pkg/operator/controller"
-	nodeclaimutil "sigs.k8s.io/karpenter/pkg/utils/nodeclaim"
 	"sigs.k8s.io/karpenter/pkg/utils/result"
 )
 
@@ -110,6 +111,45 @@ func (c *Controller) Name() string {
 	return "nodeclaim.disruption"
 }
 
+// PodEventHandler is a watcher on v1.Pods that maps Pods to NodeClaim based on the node names
+// and enqueues reconcile.Requests for the NodeClaims
+func PodEventHandler(c client.Client) handler.EventHandler {
+	return handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, o client.Object) (requests []reconcile.Request) {
+		if name := o.(*v1.Pod).Spec.NodeName; name != "" {
+			node := &v1.Node{}
+			if err := c.Get(ctx, types.NamespacedName{Name: name}, node); err != nil {
+				return []reconcile.Request{}
+			}
+			nodeClaimList := &v1beta1.NodeClaimList{}
+			if err := c.List(ctx, nodeClaimList, client.MatchingFields{"status.providerID": node.Spec.ProviderID}); err != nil {
+				return []reconcile.Request{}
+			}
+			return lo.Map(nodeClaimList.Items, func(n v1beta1.NodeClaim, _ int) reconcile.Request {
+				return reconcile.Request{
+					NamespacedName: client.ObjectKeyFromObject(&n),
+				}
+			})
+		}
+		return requests
+	})
+}
+
+// NodePoolEventHandler is a watcher on v1beta1.NodeClaim that maps Provisioner to NodeClaims based
+// on the v1beta1.NodePoolLabelKey and enqueues reconcile.Requests for the NodeClaim
+func NodePoolEventHandler(c client.Client) handler.EventHandler {
+	return handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, o client.Object) (requests []reconcile.Request) {
+		nodeClaimList := &v1beta1.NodeClaimList{}
+		if err := c.List(ctx, nodeClaimList, client.MatchingLabels(map[string]string{v1beta1.NodePoolLabelKey: o.GetName()})); err != nil {
+			return requests
+		}
+		return lo.Map(nodeClaimList.Items, func(n v1beta1.NodeClaim, _ int) reconcile.Request {
+			return reconcile.Request{
+				NamespacedName: client.ObjectKeyFromObject(&n),
+			}
+		})
+	})
+}
+
 func (c *Controller) Builder(_ context.Context, m manager.Manager) operatorcontroller.Builder {
 	return operatorcontroller.Adapt(controllerruntime.
 		NewControllerManagedBy(m).
@@ -117,11 +157,11 @@ func (c *Controller) Builder(_ context.Context, m manager.Manager) operatorcontr
 		WithOptions(controller.Options{MaxConcurrentReconciles: 10}).
 		Watches(
 			&v1beta1.NodePool{},
-			nodeclaimutil.NodePoolEventHandler(c.kubeClient),
+			NodePoolEventHandler(c.kubeClient),
 		).
 		Watches(
 			&v1.Pod{},
-			nodeclaimutil.PodEventHandler(c.kubeClient),
+			PodEventHandler(c.kubeClient),
 		),
 	)
 }
