+++ pkg/utils/nodeclaim/nodeclaim.go
@@ -24,7 +24,6 @@ import (
 	"github.com/samber/lo"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"knative.dev/pkg/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
@@ -34,29 +33,6 @@ import (
 	"sigs.k8s.io/karpenter/pkg/scheduling"
 )
 
-// PodEventHandler is a watcher on v1.Pods that maps Pods to NodeClaim based on the node names
-// and enqueues reconcile.Requests for the NodeClaims
-func PodEventHandler(c client.Client) handler.EventHandler {
-	return handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, o client.Object) (requests []reconcile.Request) {
-		if name := o.(*v1.Pod).Spec.NodeName; name != "" {
-			node := &v1.Node{}
-			if err := c.Get(ctx, types.NamespacedName{Name: name}, node); err != nil {
-				return []reconcile.Request{}
-			}
-			nodeClaimList := &v1beta1.NodeClaimList{}
-			if err := c.List(ctx, nodeClaimList, client.MatchingFields{"status.providerID": node.Spec.ProviderID}); err != nil {
-				return []reconcile.Request{}
-			}
-			return lo.Map(nodeClaimList.Items, func(n v1beta1.NodeClaim, _ int) reconcile.Request {
-				return reconcile.Request{
-					NamespacedName: client.ObjectKeyFromObject(&n),
-				}
-			})
-		}
-		return requests
-	})
-}
-
 // NodeEventHandler is a watcher on v1.Node that maps Nodes to NodeClaims based on provider ids
 // and enqueues reconcile.Requests for the NodeClaims
 func NodeEventHandler(c client.Client) handler.EventHandler {
@@ -65,22 +41,6 @@ func NodeEventHandler(c client.Client) handler.EventHandler {
 		nodeClaimList := &v1beta1.NodeClaimList{}
 		if err := c.List(ctx, nodeClaimList, client.MatchingFields{"status.providerID": node.Spec.ProviderID}); err != nil {
 			return []reconcile.Request{}
-		}
-		return lo.Map(nodeClaimList.Items, func(n v1beta1.NodeClaim, _ int) reconcile.Request {
-			return reconcile.Request{
-				NamespacedName: client.ObjectKeyFromObject(&n),
-			}
-		})
-	})
-}
-
-// NodePoolEventHandler is a watcher on v1beta1.NodeClaim that maps Provisioner to NodeClaims based
-// on the v1beta1.NodePoolLabelKey and enqueues reconcile.Requests for the NodeClaim
-func NodePoolEventHandler(c client.Client) handler.EventHandler {
-	return handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, o client.Object) (requests []reconcile.Request) {
-		nodeClaimList := &v1beta1.NodeClaimList{}
-		if err := c.List(ctx, nodeClaimList, client.MatchingLabels(map[string]string{v1beta1.NodePoolLabelKey: o.GetName()})); err != nil {
-			return requests
 		}
 		return lo.Map(nodeClaimList.Items, func(n v1beta1.NodeClaim, _ int) reconcile.Request {
 			return reconcile.Request{
@@ -200,22 +160,6 @@ func NewFromNode(node *v1.Node) *v1beta1.NodeClaim {
 	nc.StatusConditions().MarkTrue(v1beta1.Launched)
 	nc.StatusConditions().MarkTrue(v1beta1.Registered)
 	return nc
-}
-
-func Get(ctx context.Context, c client.Client, name string) (*v1beta1.NodeClaim, error) {
-	nodeClaim := &v1beta1.NodeClaim{}
-	if err := c.Get(ctx, types.NamespacedName{Name: name}, nodeClaim); err != nil {
-		return nil, err
-	}
-	return nodeClaim, nil
-}
-
-func List(ctx context.Context, c client.Client, opts ...client.ListOption) (*v1beta1.NodeClaimList, error) {
-	nodeClaimList := &v1beta1.NodeClaimList{}
-	if err := c.List(ctx, nodeClaimList, opts...); err != nil {
-		return nil, err
-	}
-	return nodeClaimList, nil
 }
 
 func UpdateNodeOwnerReferences(nodeClaim *v1beta1.NodeClaim, node *v1.Node) *v1.Node {
